scripts: requirements-extras: Add crc-tool
crc-tool is a tool wich primary use case is to input an ELF file,
parse begin and end of section through symbols originally defined
in the linker file, and add CRC after the end address, as defined
by these symbols. This is intended to be used as a post build step
to add CRCs to the generated ELF file.

Signed-off-by: Lars Thalian Morstad <l-morstad@ti.com>
Signed-off-by: Vebjorn Myklebust <v.myklebust@ti.com>
Signed-off-by: Stoyan Bogdanov <sbogdanov@baylibre.com>
Signed-off-by: Julien Panis <jpanis@baylibre.com>